+++ src/Requests/WikiSearchRequest.php
@@ -1,0 +1,13 @@
+<?php
+
+
+namespace CTApi\Requests;
+
+
+class WikiSearchRequest
+{
+    public static function search(string $query): WikiSearchRequestBuilder
+    {
+        return (new WikiSearchRequestBuilder())->search($query);
+    }
+}